+++ controllers/utility.go
@@ -3,7 +3,7 @@ package controllers
 import (
 	"time"
 
-	capiv1alpha3 "sigs.k8s.io/cluster-api/api/v1alpha3"
+	capi "sigs.k8s.io/cluster-api/api/v1beta1"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
@@ -35,29 +35,29 @@ func requeue() reconcile.Result {
 	}
 }
 
-func getClusterCreationTimeStamp(cluster *capiv1alpha3.Cluster) time.Time {
+func getClusterCreationTimeStamp(cluster *capi.Cluster) time.Time {
 	return cluster.CreationTimestamp.UTC()
 }
 
-func deletionTimeReached(cluster *capiv1alpha3.Cluster) bool {
+func deletionTimeReached(cluster *capi.Cluster) bool {
 	return time.Now().UTC().After(getClusterCreationTimeStamp(cluster).Add(defaultTTL))
 }
 
-func deletionTime(cluster *capiv1alpha3.Cluster) int {
+func deletionTime(cluster *capi.Cluster) int {
 	return int(defaultTTL.Minutes()) - int(time.Now().UTC().Sub(getClusterCreationTimeStamp(cluster)).Minutes())
 }
 
-func deletionEventTimeReached(cluster *capiv1alpha3.Cluster) bool {
+func deletionEventTimeReached(cluster *capi.Cluster) bool {
 	return time.Now().UTC().After(getClusterCreationTimeStamp(cluster).Add(eventDefaultTTL))
 }
 
-func hasChartAnnotations(cluster *capiv1alpha3.Cluster) bool {
+func hasChartAnnotations(cluster *capi.Cluster) bool {
 	releaseName, nameOK := cluster.ObjectMeta.Annotations[helmReleaseNameAnnotation]
 	releaseNamespace, namespaceOK := cluster.ObjectMeta.Annotations[helmReleaseNamespaceAnnotation]
 	return nameOK && namespaceOK && releaseName != "" && releaseNamespace != ""
 }
 
-func getChartNamespacedName(cluster *capiv1alpha3.Cluster) client.ObjectKey {
+func getChartNamespacedName(cluster *capi.Cluster) client.ObjectKey {
 	return client.ObjectKey{
 		Name:      cluster.ObjectMeta.Annotations[helmReleaseNameAnnotation],
 		Namespace: cluster.ObjectMeta.Annotations[helmReleaseNamespaceAnnotation],